- Separate .js files for batchAlternativeFeeMap and batchFeeMap - Fix error: cannot submit batch AlternativeFeeMap - Move jquery.js to headerResources.jsp instead of footerResources.jsp

+++ src/main/java/com/duke/nurseryschool/hibernate/dao/AlternativeFeeChargeMapDAO.java
@@ -58,6 +58,13 @@ public class AlternativeFeeChargeMapDAO extends CoreDAO {
 	public void deleteAlternativeFeeMap(int paymentId, int feeId) {
 		AlternativeFeeMap alternativeFeeMap = this.getAlternativeFeeMap(
 				paymentId, feeId);
+		if (alternativeFeeMap == null)
+			return;
+
+		if (!this.transaction.isActive()) {
+			this.transaction.begin();
+		}
+
 		try {
 			this.session.delete(alternativeFeeMap);
 		}
